+++ scripty/config_wrapper.py
@@ -8,7 +8,7 @@ class ConfigWrapper:
     """
 
     @staticmethod
-    def sm_conf(project_name="maximilian", team_name="data"):
+    def sm_conf(project_name:str, team_name:str):
         """
         Simple config wrapper for using secrets manager.
         """
@@ -28,12 +28,30 @@ class ConfigWrapper:
             choices=["secret_manager", "core"],
         )
 
+        parser.add_argument(
+            "-smp",
+            "--secret_project_name",
+            default="maximilian3",
+            help="""
+                secret manager project group
+                """
+        )
+
+        parser.add_argument(
+            "-smt",
+            "--secret_team",
+            default="data",
+            help="""
+                secret manager team
+                """
+        )
+
         return parser
 
     @staticmethod
     def process_config(args):
         if args.config == "secret_manager":
-            conf = ConfigWrapper.sm_conf()
+            conf = ConfigWrapper.sm_conf(project_name=args.secret_project_name, team_name=args.secret_team)
         elif args.config == "core":
             conf = config()
 
@@ -42,5 +60,7 @@ class ConfigWrapper:
         user = conf[args.database]["user"]
         password = conf[args.database]["password"]
         port = conf[args.database]["port"]
+        batchy_server = conf["batchy"]["server"]
+        batchy_port = conf["batchy"]["port"]
 
-        return db_name, host, user, password, port
+        return db_name, host, user, password, port, batchy_server, batchy_port